exit button & repair project structure
