+++ fe/src/app/app.module.ts
@@ -9,23 +9,16 @@ import { MatProgressSpinnerModule } from '@angular/material/progress-spinner';
 import { MatFormFieldModule } from '@angular/material/form-field';
 import { MatInputModule } from '@angular/material/input';
 import { MatSelectModule } from '@angular/material/select';
-
-
-
-
 import {PagesModule} from './modules/pages/pages.module';
 import { HomeComponent } from './modules/pages/components/home/home.component';
-
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
-
-import {} from '@angular/material';
-
 import { UserModule } from './modules/user/user.module';
 import { RegFormComponent } from './modules/user/components/reg-form/reg-form.component';
 import { LogInFormComponent } from './modules/user/components/log-in-form/log-in-form.component';
 import { APIInterceptor } from './interceptors/APIInterceptor';
 import {CanActivateLoginPagesService} from './services/security/can-activate-login-pages.service';
 import {CanActivateNotLoginPagesService} from './services/security/can-activate-not-login-pages.service';
+import {NavBarModule} from './modules/navigation/nav-bar.module';
 
 const appRoutes: Routes = [
   { path: '', component: HomeComponent, canActivate: [CanActivateLoginPagesService]},
@@ -49,6 +42,7 @@ const appRoutes: Routes = [
     MatInputModule,
     MatSelectModule,
     BrowserAnimationsModule,
+    NavBarModule,
     RouterModule.forRoot(appRoutes),
   ],
   providers: [APIInterceptor, {
