+++ fe/src/app/modules/navigation/components/nav-bar.component.html
@@ -3,12 +3,5 @@
   <button mat-menu-item>My Desks</button>
   <button mat-menu-item>Create Desk</button>
 </mat-menu>
-
-<!--
-<button mat-button [matMenuTriggerFor]="menu"></button>
-<mat-menu #menu="matMenu">
-  <button mat-menu-item>Cabinet</button>
-  <button mat-menu-item>Friends</button>
-</mat-menu>
--->
+<button mat-button (click)="logOut()">log out</button>
